+++ projectData.js
@@ -1,0 +1,30 @@
+const projectData = [
+  {
+    name: "Trivia Candy",
+    description:
+      "A fun trivia game built with React using the Open Trivia Database API. It's seriously addictive. Play at your own risk!",
+    url: "https://www.keithrodriguez.me/trivia-game/",
+    gitHubUrl: "https://github.com/keithro/trivia-game",
+    image1: "/trivia-game-1.png",
+    image2: "/trivia-game-2.png",
+    watermark1: "/watermark-ribbon.png",
+    watermark2: "/watermark-brain.png",
+    altText1: "Trivia game",
+    altText2: "Trivia game",
+  },
+  {
+    name: "Pick-Up Game",
+    description:
+      "A web app for starting pick-up games and other group sports or social activities. It was built with the MERN stack, featuring full CRUD capabilities, user authentication and authorization.",
+    url: "https://www.keithrodriguez.me/pickup-client/",
+    gitHubUrl: "https://github.com/keithro/pickup-client",
+    image1: "/pick-up-1.png",
+    image2: "/pick-up-2.png",
+    watermark1: "/watermark-net.png",
+    watermark2: "/watermark-ball.png",
+    altText1: "Pick-Up Game App",
+    altText2: "Pick-Up Game App",
+  },
+];
+
+export default projectData;
